:bamboo: Split useRelativeTimeRepresent
